refactored triangles into a package
